Semicolon: Improve grid layout of posts.
I'm removing Masonry for the moment—it's causing too many conflicts and I'm not sure I like it visually anymore.
I'd like to re-add support as a theme option at a later stage, but I don't think it's important enough to spend the time on at this instant.

This also adds some JS to count where we are in the grid layout, and add a clearing class to every article that appears at the beginning of a new row.
This prevents weird breaks between the grid rows where the title text is a bit long, and ensures that the grid is as evenly spaced as possible.