Semicolon: Add proper support for Infinite Scroll.
Hide the pagination menu if IS is loaded.

Use JS to count grid rows on each IS page load and add clearing classes as appropriate.
In certain circumstances, the grid may still break, so this may require a bit more
fine-tuning, but in general it should work quite reliably so long as users don't get
really crazy with their featured content.