Semicolon: Remove star-toggle if Infinite Scroll is enabled.
While this is a nice functionality, it's confusing from a UX perspective when Infinite Scroll is enabled, mostly because you end up booted back to the front page, rather than the IS "page" you were actually on when you clicked on the star button. I spent a little bit of time wrestling with pagination & page URLs in order to determine a sane way of generating the URL for the toggle, but there doesn't seem to be a great solution right now, short of rewriting this code entirely. In the interests of "getting things done", I'm just going to disable this feature if IS is active for now.

I'll add it back later, time and energy permitting.