Semicolon: Add feature-starring as a theme option that's turned off by default.
The custom feature-starring functionality for this theme is useful when you understand what it is and what it does.
The majority of users won't, particularly since it's unexpected functionality. However, this functionality is super-useful
if you understand how to use it. Accordingly, I'm hiding the functionality by default, but leaving it as a discoverable theme
option that a user can choose to activate--or ignore--as necessary.