Semicolon: Add pull-quote styling to use some negative space in single-column views.
When Semicolon is used as a single-column theme, there can be quite a bit of empty space in
the right-hand column. Offset pullquotes are used to lend some interest to this space.

I've made this user-configurable, since it pulls the blockquote out of the flow of the text,
which might be unexpected behaviour otherwise. On smaller screens or with a sidebar, the pullquote
will just have slightly more emphasized styling than other blockquotes.