Independent Publisher: Simplify .site-branding
The original theme has a featured called "multi author", which
displays the user avatar and post publish date information on
the .site-branding container when viewing single posts.

This approach can make it confusing to users by changing the site
logo and branding with the post author details and publish date.

This commit makes sure that the site logo, title and description
are always visible in all locations, which is the common
convention for the site branding.

The 'template-parts/header-site-branding.php' template has been
removed, since it is no longer needed (simpler implementation).