Independent Publisher: Refactor post cover to allow minimal height.
This change involves replacing the post cover image with
a canvas element with the width and height attributes specified,
to allow proportional resizing behavior just like image tags.

The canvas element also allows CSS backgrounds, which makes it
perfect for situations where the background must resize proportionally
down to a minimal height, in which case it would adjust the background
image using the cover background size behavior, as opposed to stretching
the image, which is the behavior when using image tags.