Independent Publisher: Increase sidebar size, adjust responsive breakpoints.
The original theme design has a 160px sidebar, which is too       
narrow. The sidebar has been resized to have a size of 250px,
which provides enough room for the widgets.

Additionally, the site's responsive breakpoints have been
adjusted, taking into account the new sidebar size.

The content width was not changed as part of this refactor,
it still has a value of 700px.

For larger screens (that is, 1380px and above), an additional
padding has been added to the content, to make sure it remains
centered on the screen, taking advantage of the additional
screen real estate.