Franklin: Swap post format link and post tags on tablet size and smaller.
Due to a limitation with CSS floats, some JavaScript is needed to
move the post format link to the left when viewing the site on
tablet and smaller screen sizes.

This is an enhancement of the actual behavior. If JavaScript is
disabled or not present, the post format link will flow naturally
to the right.