Franklin: Set main container width in pixels instead of ems.
After testing, it turns out the em units in the media queries were not causing the
container widths to change. The `#content` and `.inner` containers had their width
specified in ems, which was the real cause of the issue.

This commit implements the change to the container widths, and reverts the changes
made in r29949.