Libretto: Remove custom password form and restyle form to account for default password form.
The custom form isn't really future-friendly, so I should probably stop trying to rewrite it. :)
Seems like this was causing the form to submit incorrectly, probably due to a core update
since it was originally written a thousand years or so ago.

See #3379. This should solve the problem, but needs testing.