Pique: Fix Infinite Scroll bug and clean up button styles a bit.
The pattern library broke my theme and made it sad, but now it's happy again
and I don't have bald spots on my head from ripping out all my hair.

Note: Infinite Scroll wasn't working because it was looking for the template
file to render in the wrong folder and/or it hates me. Fixed here:
https://github.com/Automattic/slush-theme-pattern-library/pull/86