Pique: Fix a small page template bug that breaks the whole layout.
If a page had previously used a page template that doesn't exist in this theme,
get_page_template_slug(), will, much to my delight, still return the old page
template URL, even though it may not match the directory structure we use here.

That basically breaks a bunch of stuff and makes everything sad, so we're checking
first to be sure that page template actually exists before we throw it into the post class.