Pique: Re-write loops and adjust positioning of edit link so as to ensure that we always show the correct edit link.
We have some custom loops within custom loops here, and that was causing problems and
outputting the same edit link (to the homepage) every time. Not ideal! wp_reset_postdata()
isn't super helpful in this instance, since it resets the post data to the main loop, which,
in this case, is the front page itself (rather than the individual panels). In the interests
of keeping things simple(ish) I've just rearranged things so they work properly.