Pique: Ensure priority plus navigation pattern is correctly applied.
First, we're adding a conditional to ensure that it's only added to the
primary navigation, just to avoid any messiness with the secondary menu.

Next, we've removed some selectors that were specific to the split-nav
functionality that has since been removed from Pique. These selectors were
causing the priority+ menu to not work anymore, leading to lots of ugliness
on mobile devices. Correcting the selectors makes it work properly again.

Hurray!