Gateway: Adding check in JS to make sure browser window is tall enough before actually making masthead sticky.
From what I can tell, the issue is that when the masthead sticks, it a) causes the site to scroll and b) makes the page shorter. If the page is made short enough, it's no longer possible scroll enough for the masthead to stick, so it unsticks. This makes the page taller again, so the masthead sticks again, which causes the page to be too short again...

Change causes masthead not to stick at really specific window heights.

Fixes #3421.