Pique: Rework menus a few smidges.
The menus for this are tripping me up a bit. For the moment, I've re-organised the
code so it's easier to parse & work with, and I've added carets to indicate menu items
with children. I've also improved the look & feel of the menu on mobile devices.

There are still a few issues that need to be addressed here, so it's in-progress.