Pique: Rework post_class filter so it's easier to parse & read.
My code for this was a bit hard to follow, so I've broken it up into more
pieces. Less elegant, more lines, but more readable! Also something was
broken with the demo site and I think now this fixes it. If I never have
to look at this code again I will be a very happy lady.