Pique: Improve the behaviour and display of menus on small screens.
The priority+ navigation scheme meant that the boilerplate code used in _s
really didn't work here. So I've rewritten most of it, using jQuery for ease-of-use
(and also because we're already loading it for use elsewhere in the theme.)

What's happening here now: when you click on a menu item (with sub-menus) once,
the sub-menu (and all sub-sub menus) will appear. If you click that menu item again
(and it actually has a link associated with it), it goes to the link as expected.
However, if it doesn't have a link associated with it (ie, it just leads to a hash tag)
it closes the menu. 

May still require a bit of tweaking/testing, but wayyyyyy better than what I had before.