Pique: Finally add some Customizer magic!
We need to do a bit of hocus-pocus to make this all work: basically, we need some JS on the
admin side, hooked to `customize_controls_enqueue_scripts`, that listens for when the panel
sections are opened in the controls iframe, then sends that information over to the main 
customizer.js, where we can modify the site preview iframe. This file then has some extra 
code that collects the data from the controls and uses it to scroll to the panel in question
and flash its border, providing some better feedback as to what we're actually modifying.

This change definitely didn't take hours of work. 

WHEEEEEE!