Pique: Rearrange header items on mobile devices.
Instead of setting a max-height, we're going to cut down on items shown on the header, so it can
scale responsively and still not get all overflowy. So if there's a logo we won't display the title,
and the description goes away altogether. We're also moving the navbar down as much as we can so it all
fits together a bit better. Fixes #3596.