Cerauno: Addresses issue where Pinterest and LinkedIn icons wouldnt display in the Social Media Icons widget
The wpcom styles file includes an override for The Social Media widget that replaces pinterest-alt and linkedin-alt genericons with the Font Awesome icons included in the theme. This does not account for standard icon classes and so those -alt overrides didn’t have an icon declared for the widget. 

To fix this, I made it so that the override catches non '-alt' icons as well which did the trick.

Fixes #3770