Since r52410, we're saving Contact info values in blog options rather than theme_mods.
For blogs created prior to this change, we created at patch to make sure we fall back on the value from theme_mod, where available, in the Customizer.

A bug report brought to our attention a situation in which a user wished to empty Contact info values via the Customizer. Because of the fall back check, if the theme_mod value is present it will always display and never update.

This patch introduces custom sanitize methods, which serve as a way for us to remove these theme_mod values altogether.