Update Varia and Maywood to include FSE fixes and support
Summary:
This includes a number of fixes for Full Site Editing support, and also whitelists Maywood to support Full Site Editing flows.

If this is tested and lands, let's also make sure we make this available with Jetpack and Atomic sites. We may need to ask folks for help on steps.

Test Plan: Verify no regressions for Maywood and Varia on Non Full Site Editing Sites

Reviewers: #cylon_team, #serenity_team, apeatling, iamtakashi, davidalankennedy, helpingcat

Reviewed By: apeatling

Subscribers: copons

Differential Revision: D32421