P2: redesign private.php landing page if not member of a P2 site or logged out v2
Summary:
Addresses https://github.com/Automattic/p2/issues/287.

In this patch, we redesign the landing page of a P2 site if it's visibed by a logged out user or a non-member (private.php).

Partly a deploy of https://github.com/Automattic/p2/pull/785.

Test Plan:
Apply D48684 in `public_html/` and this patch in `wp-content/themes/pub`.

Sandbox a P2 site (like lamosty.p2.blog). Open the site in an incognito window. You should be presented with the "This P2 is private" and the P2 branding (like logo). You should see a button to log in. Now, log into a non-a8c account which is not a member of that site. You should see "You need permission to access and participate in this P2." and a button to request access to the P2 site. Clicking it should work.

If a P2 is an a8c one and you are not a member yet, you should see a message about giving yourself an access to the P2 site.

Reviewers: #lighthouse_team

Subscribers: aidvu, evilluendas, vortfu

Differential Revision: D47616