#2415 Coutoire: Added missing bold font on headings
Summary:
Added the bold font weight inside functions.php coutoire_fonts_url function because we only had 400 and 400i for Garamond.

I also had to default the headings to a weight of 400 for the bolder rule to pick up the correct weight.

Test Plan:
Create a new heading and make it bold. Check if it's bold on the front-end




Reviewers: #themes_team

Differential Revision: D49726