Fix/varia children pullquote color issue #2502
Summary: Changed the inheritance of the pullquote block background color so that the selected color from the editor is more specific. This affects Varia and all its children

Test Plan: Create a new Pullquote block, make it solid color, change the main color and see if the frontend has the same color as the editor.

Reviewers: #themes_team

Differential Revision: D50130