Seedlet: fix for mobile scroll on alignfull galleries #2471
Summary: The margin value for galleries was affecting the responsive style for full width galleries and causing a scroll. I omitted the margin in the case where the gallery is full width so that the .alignfull rules will apply to them.

Test Plan: Create a full width gallery and check that you can't scroll horizontally on mobile. The gallery should be full width.

Reviewers: #themes_team

Differential Revision: D50493