Gazette: changed height to post-thumbnail #2527
Summary: Images of posts in a 3 columns recent posts block have a black area underneath them. This is caused by an incorrect height value on the post-thumbnail wrapper.

Test Plan: Create a page with a block of recent posts that have featured images. There should be no black area under the images

Reviewers: #themes_team

Differential Revision: D50615