Fix editor styles for verse block for all Varia child themes
Summary:
Update varia child themes. Copied from https://github.com/Automattic/themes/pull/2640

Basically, make sure the verse block uses the correct font style and font color in the editor as it already does on the front end.

Test Plan:
Test on various varia child themes. Specifically, the weird ones like Coutoire, Rivington, etc. Make sure light-on-dark text works, and that it is always monospace.

Note that it's not a different font because the front end is already monospace all the time, so we shouldn't change that.

basically, to test:
1. sandbox a test site
2. apply these changes to wp-content/themes/pub
3. activate various varia child themes
4. compare a post with a verse block in the editor to the front end. Make sure font and color is more or less the same.

Reviewers: tomjcafferkey

Reviewed By: tomjcafferkey

Subscribers: tomjcafferkey

Differential Revision: D51536