Added header padding when there is no menu
Summary:
There was no padding value assigned to the header when there was not a
menu available. This change uses the same padding values as are used
WITH a menu for when there is none.

Test Plan:
Install Spearhead theme.
Setup a menu and assign it to 'Primary Navigation'
Note the location of the title (in the header). (This is the expected location)
Remove the 'Primary Navigation' menu (via the customizer)
Note the location of the title.
The expected location is the same location as when there is a 'Primary
Navigation' menu assigned.

Reviewers: #themes_team, jacklenox

Reviewed By: #themes_team, jacklenox

Subscribers: jacklenox

Differential Revision: D51939