Morden and Varia: fixed conditional class not working corectly #2798
Summary: Fixed condition that was never being met for the headers with tagline + site title.

Test Plan:
On a site with Morden, Hever or Varia themes add a logo, site title and tagline.
Check that the customizer and the frontend look the same.

Related issue: https://github.com/Automattic/themes/issues/1467

Reviewers: #themes_team

Differential Revision: D52560