Responsive: Gallery Widget Unresponsive
Summary:
The IE6 upgrade check (now obsolete enough to eliminate) depended on jquery.browser to determine if an upgrade should be prompted.
Since jQuery has been updated and now no longer provides jQuery.browser the script threw a null reference error and stopped executing the remainder of the script.
The offending ie6 test script leveraging this code has been removed.

Test Plan:
Only here have I been able to reproduce: https://andrechouraqui.com/
Load this page (using Chrome or some other browser that allows for the editing of a live site).
Note the null ref error thrown in the console and navigate to that code block.
Pause JS execution at that line, comment that line and the following (if (jquery.browser...)
Note that the side gallery widgets load as expected.

Note that those two lines are the two removed in this DIFF.

Reviewers: #themes_team, scruffian

Reviewed By: #themes_team, scruffian

Subscribers: scruffian

Differential Revision: D52184