fixed z-index of header semicircle
Summary: Just changing z-index of the ::before pseudo-element was causing problems with the semicirle overlapping the site title so I decided to move it from #page to #masthead since that's where the other semicircle is.

Test Plan:
Activate Brompton on your test site. Select to use Brompton's home page.
Go to edit homepage
Make sure that the cover image has an overlay active
Check the frontend, both semi circles should be on top of the red borders of the header and the whole header should be on top of the cover below.

Reviewers: #themes_team

Differential Revision: D52682