TT1-Blocks: Add full width alignment for block template post content
Summary:
**Description:**

Post content blocks in the tt1-blocks theme were defaulting to a normal width alignment on a fresh install.  It appeared as if the post content children weren't responding changes to "full" and "wide" alignment, when in reality, it was their parent, the post content block, that was limiting the alignment changes.

As a result, we modify the post-content block templates to default to "full" width for all theme provided tt1 post content blocks.

**Note:**
See https://github.com/WordPress/theme-experiments/pull/205 for more context.

Test Plan:
1. Create a new full site editing enabled site
2. Select the tt1-blocks theme
3. Verify that the post content blocks in the page-home, single, index, and page templates render with a default, normal alignment
4. Sandbox site and the public api
5. Verify that the post content blocks in the page-home, single, index, and page templates render with a full alignment

Reviewers: #cylon_team, #themes_team, kjellreigstad, scruffian, vindl

Reviewed By: #cylon_team, #themes_team, scruffian, vindl

Subscribers: vindl

Differential Revision: D57025