Add _hs_front_page meta for Spearhead annotations.
Summary:
Spearhead currently does not work correctly in the theme switching flow to selecting using the themes homepage starter content.

This seems to be the result of no pages having the `_hs_front_page` custom meta.

Currently I have added this to the "Home" page, although this pages post content looks fairly different from the preview in the theme thumbnail.  Maybe this homepage content needs updated, or maybe we need to be setting something else as the homepage or preview? {U+01F914} 

Test Plan:
* apply this diff
* switch to spearhead theme and select to use the new themes homepage content.
* verify homepage content changes.

Reviewers: kjellreigstad, themeshaper, ramonopoly

Reviewed By: ramonopoly

Subscribers: ramonopoly

Differential Revision: D57031