Seedlet: Fix fatal errors when previewing Seedlet
Summary:
This fixes a fatal error caused by previewing Seedlet, Spearhead and Blank Canvas in the Customizer when running a different theme.

The problem comes about because passing a theme parameter to the Customizer is overriding the theme, but not updating whatever get_template() returns.

Strictly speaking this is an issue with theme previews in the Customizer, but I don't believe its worth our time and effort to fix this feature. 

If we were only thinking about wpcom we could simply hardcode Seedet's path in to the require, but that won't work on Atomic.

Instead we can just duplicate the code, since we don't plan on making any more Seedlet child themes.





Test Plan:
- Switch to any other theme than Seedlet
- Go to the Theme Showcase https://wordpress.com/themes/
- Find Seedlet and click "Live Demo"
- Then click "Try & Customize"
- You should see color palletes in the customizer
- You should not see any fatal errors in your logs
- Change the theme preview to spearhead by updating the URL to `pub/spearhead`
- Check for fatals again
- Change the theme preview to blank-canvas by updating the URL to `pub/blank-canvas`
- Check for fatals again

Reviewers: #themes_team, jffng

Reviewed By: #themes_team, jffng

Subscribers: jffng

Differential Revision: D59314