Dara: Update annotations to better match the demo site.
Summary:
This is a tricky one, because the demo has a lot of bells and whistles that are hard to set up from scratch (Featured Content, Dara Featured Pages, Testimonials, widgets), but that means there's a lot of content to overwrite in the annotation. I wasn't able to get the Dara Featured Pages section to work using the annotation, either.

Before

{F8786938}

After

{F8786951}

Test Plan:
* Apply this patch to your sandbox in the `/wp-content/themes/pub` repo and sandbox your test site
* You may need to manually link Phabricator to this repo by running `arc set-config default https://code.a8c.com` in this directory
* Run Headstart on your test site: `wp headstart run --url=yourtestsite.wordpress.com --user=yourusername--theme=pub/dara --delete-content`
* The result should look something like the "After" screenshot above.

Reviewers: #serenity_team, themeshaper, mreishus

Reviewed By: #serenity_team, mreishus

Subscribers: mreishus, blackjacksoftware

Differential Revision: D61610