Add support for Editing Toolkit Font Picker to 2021
Summary: Because the font system was already prepaired for it only enabling the font support was needed to enable the Editing Toolkit Font Picker.

Test Plan:
Activiate TwentyTwentyOne on a .com simple site.  
Edit a post.  
Note that the 'Global Styles' Font picker is available.  
Change the fonts and save.  
Note that the changed fonts are also visible in the view.

Reviewers: #themes_team, jffng

Reviewed By: #themes_team, jffng

Subscribers: jffng

Differential Revision: D62156