Remove errant float from comment elements from 2019 theme
Summary:
A float was inappropriately applied to comment form.  This change removes
the float and recompiles the resources

Test Plan:
Setup a wordpress.com site that accepts comments from unauthenticated users.
Browse to a page allowing comments in an icognito or unauthenticated browser.
Attempt to add a comment.
Observe how correct the form looks.

Reviewers: #themes_team, scruffian

Subscribers: scruffian

Differential Revision: D62215