Theme Varia and children: Fix font color on cover blocks
Summary:
When you add a cover block with a white (or clear) background the font color is also white and you cannot see what you write.
This change makes the necessary changes so that the cover font colors are picked up from the customizer palette in Colors & Backgrounds > Text Color.
This fix applies to Varia and all its child themes.

The changes are in the sass files in the Varia template:
* varia/sass/blocks/utilities/_style.scss
* varia/sass/blocks/utilities/_editor.scss
* varia/sass/blocks/cover/_style.scss
* varia/sass/blocks/cover/_editor.scss

All the other changes are auto-generated following this instructions on theme p2.

After/if merging there are still instructions to be followed in the P2 mentioned above.

Before and after images are available in this PR: https://github.com/Automattic/themes/pull/4086

Test Plan:
Verify on your sandbox that when you add a cover block with a light background-color the paragraph inside is visible. This only applies to Varia and its child themes. Here is a list of the themes affected:

```
Barnsbury
Dalston
Mayland
Rivington
Maywood
Balasana
Shawburn
Alves
Exford
Rockfield
Stratford
Coutoire
Morden
Stow
Hever
Leven
Varia
```

Reviewers: #serenity_team, sixhours

Reviewed By: #serenity_team, sixhours

Subscribers: sixhours

Tags: #serenity_team

Differential Revision: D63216