Independent Publisher 2: Make color annotations button rule more specific
Summary:
`button:not('.menu-toggle')` was meant to avoid applying color annotation styles to the Menu toggle on mobile devices, but it has the unpleasant side effect of targeting all buttons that are not `.menu-toggle`, even if you're in the editor.

Removing the `:not('.menu-toggle')` specificity seems to work fine in the color annotations without changing things for existing users, and should fix the editor.

**Before**

{F8966702}

**After**

{F8966703}

Test Plan:
* Apply this patch to your sandbox, then grab the Independent Publisher 2 theme
* Rename the theme directory and change the name in `style.css` to `Independent Publisher Test` and ZIP it.
* Switch to an Atomic site and upload the zipped theme, and activate it
* Go to the Customizer and set the color scheme on your Atomic site such that the third color circle is black (or any color you'll notice)
* Save your changes
* Go to the editor by opening a page or post
* The editor buttons should all appear normal, without a black background.

Reviewers: mreishus, #themes_team

Reviewed By: mreishus

Differential Revision: D65100