Add blank-canvas-blocks
Summary:
We're renaming `blank-canvas-2` to `blank-canvas-blocks` because we want to settle on a common slug for all FSE versions of themes. We've decided to add a trailing `-blocks suffix`.

But first, we need to publish the theme.

Test Plan: No testing is required. The theme is not published yet.

Reviewers: #cylon_team, jeremyyip

Reviewed By: #cylon_team, jeremyyip

Subscribers: jeremyyip

Tags: #cylon_team